requirements.txt
aiogram==3.15.0
python-decouple==3.8
